define IncompatibleSchemaError in dbt/artifacts/exceptions
