+++ core/dbt/artifacts/exceptions/__init__.py
@@ -1,0 +1,1 @@
+from dbt.artifacts.exceptions.schemas import IncompatibleSchemaError
