+++ core/dbt/contracts/state.py
@@ -2,12 +2,12 @@ from pathlib import Path
 from typing import Optional
 
 from dbt.contracts.graph.manifest import Manifest
+from dbt.artifacts.exceptions import IncompatibleSchemaError
 from dbt.artifacts.schemas.manifest import WritableManifest
 from dbt.artifacts.schemas.freshness import FreshnessExecutionResultArtifact
 from dbt.artifacts.schemas.run import RunResultsArtifact
 from dbt_common.events.functions import fire_event
 from dbt.events.types import WarnStateTargetEqual
-from dbt.exceptions import IncompatibleSchemaError
 
 
 def load_result_state(results_path) -> Optional[RunResultsArtifact]:
