+++ tests/functional/artifacts/test_previous_version_state.py
@@ -7,7 +7,7 @@ import pytest
 from dbt.artifacts.schemas.base import get_artifact_schema_version
 from dbt.contracts.graph.manifest import WritableManifest
 from dbt.artifacts.schemas.run import RunResultsArtifact
-from dbt.exceptions import IncompatibleSchemaError
+from dbt.artifacts.exceptions import IncompatibleSchemaError
 from dbt.tests.util import run_dbt, get_manifest
 
 # This project must have one of each kind of node type, plus disabled versions, for
